requirements.txt
requests
playwright
pyautogui
mysql-connector-python